@@ -1,49 +1,58 @@
 {
-  "name": "n",
-  "version": "0.1.0",
+  "name": "calendar",
   "private": true,
+  "version": "0.0.0",
+  "type": "module",
+  "scripts": {
+    "dev": "vite",
+    "build": "tsc -b && vite build",
+    "lint": "eslint . --ext ts,tsx --report-unused-disable-directives --max-warnings 0",
+    "server": "nodemon --exec tsx server/index.ts",
+    "preview": "vite preview"
+  },
   "dependencies": {
-    "@testing-library/jest-dom": "^5.17.0",
-    "@testing-library/react": "^13.4.0",
-    "@testing-library/user-event": "^13.5.0",
-    "@types/jest": "^27.5.2",
-    "@types/node": "^16.18.100",
-    "@types/react": "^18.3.3",
-    "@types/react-dom": "^18.3.0",
-    "google-auth-library": "^9.11.0",
+    "@types/http-proxy-middleware": "^1.0.0",
+    "bcrypt": "^5.1.1",
+    "connect-mongodb-session": "^5.0.0",
+    "cookie-parser": "^1.4.6",
+    "cookie-session": "^2.1.0",
+    "cors": "^2.8.5",
+    "dotenv": "^16.4.5",
+    "esm": "^3.2.25",
+    "express": "^4.19.2",
+    "express-http-proxy": "^2.0.0",
+    "express-session": "^1.18.0",
+    "http-proxy": "^1.18.1",
     "http-proxy-middleware": "^3.0.0",
+    "mongodb": "^6.8.0",
+    "nodemon": "^3.1.4",
+    "passport": "^0.7.0",
+    "passport-google": "^0.3.0",
+    "passport-google-oauth20": "^2.0.0",
+    "path": "^0.12.7",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
     "react-helmet-async": "^2.0.5",
     "react-router-dom": "^6.23.1",
-    "react-scripts": "5.0.1",
-    "typescript": "^5.0.4",
-    "web-vitals": "^2.1.4"
+    "sass": "^1.77.6",
+    "tsx": "^4.16.0",
+    "url": "^0.11.3",
+    "vite-plugin-pwa": "^0.20.0"
   },
-  "scripts": {
-    "dev": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject",
-    "server": "ts-node --esm server/index.ts",
-    "tsc": "tsc --version"
+  "devDependencies": {
+    "@types/node": "^20.14.8",
+    "@types/react": "^18.3.3",
+    "@types/react-dom": "^18.3.0",
+    "@typescript-eslint/eslint-plugin": "^7.13.1",
+    "@typescript-eslint/parser": "^7.13.1",
+    "@vitejs/plugin-react-swc": "^3.5.0",
+    "eslint": "^8.57.0",
+    "eslint-plugin-react-hooks": "^4.6.2",
+    "eslint-plugin-react-refresh": "^0.4.7",
+    "ts-node": "^10.9.2",
+    "tslib": "^2.6.3",
+    "typescript": "^5.2.2",
+    "vite": "^5.3.1"
   },
-  "eslintConfig": {
-    "extends": [
-      "react-app",
-      "react-app/jest"
-    ]
-  },
-  "browserslist": {
-    "production": [
-      ">0.2%",
-      "not dead",
-      "not op_mini all"
-    ],
-    "development": [
-      "last 1 chrome version",
-      "last 1 firefox version",
-      "last 1 safari version"
-    ]
-  }
+  "module": "commonjs"
 }
