@@ -1,6 +1,6 @@
 import dotenv from "dotenv";
 
-dotenv.config({ path: './credentials.env.local' });
+dotenv.config({ path: './server/modules/credentials.env.local' });
 
 /**
  * The URI for the MongoDB database.
@@ -35,8 +35,19 @@ const CLIENT_PORT: number = parseInt(process.env.CLIENT_PORT!) || 3000;
 /**
  * The port number for the server.
  */
-const SERVER_PORT: number = parseInt(process.env.SERVER_PORT!) || 3001;
+const SERVER_PORT: number = process.env.PORT as unknown as number || parseInt(process.env.SERVER_PORT!) || 3001;
 
+const env: object = {
+    URI,
+    CLIENT_DB,
+    CLIENT_ID,
+    CLIENT_SECRET,
+    APP_HOSTNAME,
+    CLIENT_PORT,
+    SERVER_PORT,
+}
+
+export default env;
 export {
     URI,
     CLIENT_DB,
